Fixed a header file

+++ PolygonChecker/main.h
@@ -1,4 +1,4 @@
 #pragma once
 void printWelcome();
 int printShapeMenu();
-int* getTriangleSides(float* triangleSides);
+float* getTriangleSides(float* triangleSides);
